pyproject.toml
[project]
name = "prasad"
version = "0.1.0"
description = ""
authors = [
    {name = "robiulsunnyemon",email = "robiulsunyemon@gmail.com"}
]

requires-python = "^3.11"
dependencies = [
    "poetry-core (>=2.0.0)",
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "beanie (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "prasad", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
